modularize the libcore/types code
