+++ src/libcore/signTransaction.js
@@ -2,7 +2,7 @@
 
 import Transport from "@ledgerhq/hw-transport";
 import type { CryptoCurrency, Account, TokenAccount } from "../types";
-import type { CoreCurrency, CoreEthereumLikeTransaction } from "./types";
+import type { CoreCurrency } from "./types";
 import byFamily from "../generated/libcore-hw-signTransaction";
 
 export default (opts: {
@@ -12,5 +12,5 @@ export default (opts: {
   tokenAccount: ?TokenAccount,
   currency: CryptoCurrency,
   coreCurrency: CoreCurrency,
-  coreTransaction: CoreEthereumLikeTransaction
+  coreTransaction: *
 }) => byFamily[opts.account.currency.family](opts);
